refactor: move analytics into vertical feature
